+++ src/main/java/br/todolist/todoapp/controller/TaskController.java
@@ -4,9 +4,10 @@ import org.springframework.web.bind.annotation.RestController;
 
 import br.todolist.todoapp.model.Task;
 import br.todolist.todoapp.repository.TaskRepository;
+import br.todolist.todoapp.security.services.JWTUtils;
 import jakarta.validation.Valid;
-import jakarta.validation.ValidationException;
 
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -17,6 +18,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.MethodArgumentNotValidException;
+import org.springframework.web.bind.annotation.CookieValue;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -28,59 +30,87 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 
 @RestController
 public class TaskController {
-    @Autowired TaskRepository repository;
+    @Autowired
+    TaskRepository repository;
+    @Autowired
+    JWTUtils jwtUtils;
 
     Logger logger = LoggerFactory.getLogger(TaskController.class);
 
-    @GetMapping("/tasks")
-    public ResponseEntity<List<Task>> getTasks() {
-       List<Task> tasks = repository.getAll();
-    
-        return new ResponseEntity<>(tasks, HttpStatus.OK);
-    }
-
     @GetMapping("/tasks/{id}")
-    public ResponseEntity<List<Task>> getTask(@PathVariable(required = true) int id) {
+    public ResponseEntity<List<Task>> getTask(@PathVariable(required = true) int id,
+            @CookieValue("session") String session) {
+        var user = jwtUtils.decodePayloadToMap(session);
+        var userId = Integer.parseInt(user.get("id"));
+
         List<Task> task = repository.get(id);
 
-        if(task.isEmpty()) {
+        if (task.isEmpty()) {
             return new ResponseEntity<>(task, HttpStatus.NOT_FOUND);
         }
-    
+
+        if (task.get(0).getClientId() != userId) {
+            return new ResponseEntity<>(new ArrayList<>(), HttpStatus.UNAUTHORIZED);
+        }
+
         return new ResponseEntity<>(task, HttpStatus.OK);
     }
 
     @ResponseStatus(value = HttpStatus.NO_CONTENT)
     @PostMapping("/tasks")
     public void save(@RequestBody @Valid Task task) {
-        logger.info("Create Todo: " + task.toString());
-
         repository.save(task);
     }
 
-    @ResponseStatus(value = HttpStatus.NO_CONTENT)
     @PutMapping("/tasks/{id}")
-    public void update(@PathVariable(required = true) int id, @RequestBody @Valid Task task) {
+    public ResponseEntity<Void> update(@PathVariable(required = true) int id, @RequestBody @Valid Task task,
+            @CookieValue("session") String session) {
+        var user = jwtUtils.decodePayloadToMap(session);
+        var userId = Integer.parseInt(user.get("id"));
+
+        List<Task> t = repository.get(id);
+
+        if (t.isEmpty()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
+        if (t.get(0).getClientId() != userId) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+
         repository.update(id, task);
+
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 
-    @ResponseStatus(value = HttpStatus.NO_CONTENT)
     @DeleteMapping("/tasks/{id}")
-    public void delete(@PathVariable(required = true) int id) {
+    public ResponseEntity<Void> delete(@PathVariable(required = true) int id, @CookieValue("session") String session) {
+        var user = jwtUtils.decodePayloadToMap(session);
+        var userId = Integer.parseInt(user.get("id"));
+
+        List<Task> t = repository.get(id);
+
+        if (t.isEmpty()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
+        if (t.get(0).getClientId() != userId) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+
         repository.delete(id);
+
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 
     @ExceptionHandler(MethodArgumentNotValidException.class)
     public ResponseEntity<Map<String, String>> handler(MethodArgumentNotValidException error) {
         Map<String, String> response = new HashMap<>();
 
-        error.getFieldErrors().forEach(e -> 
-            response.put(e.getField(), e.getDefaultMessage())
-        );
+        error.getFieldErrors().forEach(e -> response.put(e.getField(), e.getDefaultMessage()));
 
         return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
     }
-    
 
     @ExceptionHandler(Exception.class)
     public ResponseEntity<Map<String, String>> handler(Exception error) {
